# [SPARK-31553][SQL] Fix isInCollection for collection sizes above the optimisation threshold