--- a/sql/core/src/test/scala/org/apache/spark/sql/ColumnExpressionSuite.scala
+++ b/sql/core/src/test/scala/org/apache/spark/sql/ColumnExpressionSuite.scala
@@ -870,10 +870,14 @@ class ColumnExpressionSuite extends QueryTest with SharedSparkSession {
       Row(Row("a", 2, 1.0)) :: Nil)
   }
 
-  test("fix in set") {
-    val set = (0 to 20).map(_.toString).toSet
-    val data = Seq("1").toDF("x")
-    assert(set.contains("1"))
-    checkAnswer(data.select($"x".isInCollection(set)), Row(true))
+  test("SPARK-31553: isInCollection for collection sizes above a threshold") {
+    val threshold = 100
+    withSQLConf(SQLConf.OPTIMIZER_INSET_CONVERSION_THRESHOLD.key -> threshold.toString) {
+      val set = (0 until 2 * threshold).map(_.toString).toSet
+      val elem = "10"
+      val data = Seq(elem).toDF("x")
+      assert(set.contains(elem))
+      checkAnswer(data.select($"x".isInCollection(set)), Row(true))
+    }
   }
 }