### What changes were proposed in this pull request?
The `InSet` expression expects input collections of internal Catalyst types, for example `hset` must contain elements of `UTF8String` for `child` of string type. So, it means `isInCollection` must convert users values to internal Catalyst values but currently it doesn't perform the conversion. That leads to incorrect results for collection sizes above the threshold `spark.sql.optimizer.inSetConversionThreshold`.

The bug was introduced by https://github.com/apache/spark/pull/25754.

### Why are the changes needed?
The changes fix incorrect behaviour of `isInCollection`. For example, if the SQL config `spark.sql.optimizer.inSetConversionThreshold`is set to 10 (by default):
```scala
val set = (0 to 20).map(_.toString).toSet
val data = Seq("1").toDF("x")
data.select($"x".isInCollection(set).as("isInCollection")).show()
```
The function must return **'true'** because "1" is in the set of "0" ... "20" but it returns "false":
```
+--------------+
|isInCollection|
+--------------+
|         false|
+--------------+
```

### Does this PR introduce any user-facing change?
Yes

### How was this patch tested?
- By existing test suite `ColumnExpressionSuite`
- Add new test to `ColumnExpressionSuite`
- Manually benchmark it by the code from #25754:
```scala
def testExplainTime(collectionSize: Int) = {
        val df = spark.range(10).withColumn("id2", col("id") + 1)
        val list = Range(0, collectionSize).toList
        val startTime = System.currentTimeMillis() 
        df.where(col("id").isInCollection(list)).where(col("id2").isInCollection(list)).explain()
        val elapsedTime = System.currentTimeMillis() - startTime
        println(s"cost time: ${elapsedTime}ms")
}
```
Then test on collection size 5, 10, 100, 1000, 10000, test result is:

collection size | explain time (before) | explain time (after) | w/o optimization
-- | -- | -- | --
5 | 64ms | 65ms | 62ms
10 | 68ms | 64ms | 88ms
100 | 41ms | 162ms | 227ms
1000 | 98ms | 406ms | 652ms
10000 | 654ms | 2579ms | 4504ms