@MaxGekk 2020-04-24
The build https://github.com/apache/spark/pull/28328#issuecomment-618992335 fails on
 org.apache.spark.sql.hive.thriftserver.CliSuite.SPARK-11188 Analysis error reporting
I don't think it is related to the changes