@dongjoon-hyun 2020-04-24
This is a nice catch, @MaxGekk . As I wrote in the comment, it would be great if we can proceed two PR separately. Thanks!