@dongjoon-hyun 2020-04-25
I merged https://github.com/apache/spark/pull/28343 first. Could you rebase this PR, @MaxGekk ? Thanks.