@MaxGekk 2020-04-25
@dongjoon-hyun I have rebased but the test from this PR starts failing. The function `convertToScala(elem, child.dataType)` doesn't convert `UTF8String` to `String` because `child.dataType` is `NullType` when InSet is created from `isInCollection`.

The test passed here because I don't wrap the result of `convertToScala` by Literal.

Probably, we need to revert https://github.com/apache/spark/commit/7d8216a6642f40af0d1b623129b1d5f4c86bec68 . Sorry, my bad.