@MaxGekk 2020-04-25
@dongjoon-hyun @cloud-fan Regarding to this PR, WDYT of reverting the optimization https://github.com/apache/spark/pull/25754 instead of to fix it by this PR?