@MaxGekk 2020-04-25
I have fixed the test failure after rebasing on https://github.com/apache/spark/pull/28343 by passing element type from the place where the type is known https://github.com/apache/spark/pull/28328/commits/67f34a10a6bafe3a7da9480fc3f79cd041cb12b0 . I could open a follow-up PR for #28343 @dongjoon-hyun Let me know if you are ok with that.