@MaxGekk 2020-04-26
Failures of CliSuite annoy a lot https://github.com/apache/spark/pull/28328#issuecomment-619520623 @dongjoon-hyun @cloud-fan @gatorsmile This is the PR to repeat tests from CliSuite https://github.com/apache/spark/pull/28329