@cloud-fan 2020-04-27
cc @viirya , this is another instance that merging `InSet` and `Set` can fix the issue.