@viirya 2020-04-27
> Actually this PR shows we still need `InSet`, to make the analyzer fast...

What that means? We optimize `In` with `InSet` in optimizer, right?