@cloud-fan 2020-04-27
A `In` with many values is slow to analyze, as the type coercion rules or `In.resolved` are very slow.