@MaxGekk 2020-04-27
I have update PR's description and added a column w/o optimization. I got the numbers by running the code:
```scala
  test("isInCollection benchmark") {
    def testExplainTime(collectionSize: Int) = {
      val df = spark.range(10).withColumn("id2", col("id") + 1)
      val list = Range(0, collectionSize).toList
      val startTime = System.currentTimeMillis()
      df.where(col("id").isInCollection(list)).where(col("id2").isInCollection(list)).explain()
      val elapsedTime = System.currentTimeMillis() - startTime
      println(s"cost time: ${elapsedTime}ms")
    }
    withSQLConf(SQLConf.OPTIMIZER_INSET_CONVERSION_THRESHOLD.key -> "100000000") {
      testExplainTime(1)
      testExplainTime(5)
      testExplainTime(10)
      testExplainTime(100)
      testExplainTime(1000)
      testExplainTime(10000)
    }
  }
```