@viirya 2020-04-27
> A `In` with many values is slow to analyze, as the type coercion rules or `In.resolved` are very slow.

That's a pain point. But when we merge `In` and `InSet`, we can have a constructor similar to current `InSet`? We can have both `list` and `hset` in `In`.  For example, in `isInCollection`, we still can construct `In` which has empty children but only `hset`?