@HyukjinKwon 2020-04-28
Okay, the more I look, the more it makes me to think we should revert #25754 rather than adding bandaid fixes. Shall we revert?