@MaxGekk 2020-04-28
After offline discussion with @gatorsmile @cloud-fan @HyukjinKwon, we decided to revert https://github.com/apache/spark/pull/25754 . I will open a PR for that and close this PR.